# docs/API_설계서/api.yaml
openapi: 3.0.0
info:
  title: 공고 정보 조회 API
  version: v1.0.0
  description: 공고 목록과 채팅 기능을 제공하는 API.

paths:
  # 채팅 메시지 등록 및 AI 응답
  /api/chat:
    post:
      summary: 사용자 - 신규 채팅 메시지 등록 및 AI 응답 받기
      operationId: postChatMessage
      tags:
        - 채팅
      requestBody:
        description: 사용자 메시지 정보
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: 성공적으로 메시지를 등록하고 AI 응답을 받았습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: 요청 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # 채팅 히스토리 목록 조회
  /api/chathistories:
    get:
      summary: 사용자 - 채팅 히스토리 목록 조회
      operationId: getChatHistories
      tags:
        - 채팅 히스토리
      parameters:
        - name: user_key
          in: query
          description: 사용자 키
          required: True
          schema:
            type: string
      responses:
        "200":
          description: 성공적으로 채팅 히스토리 목록을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoriesResponse"
        "400":
          description: 요청 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # 특정 채팅 히스토리 상세 조회
  /api/chathistories/{session_key}:
    get:
      summary: 사용자 - 특정 히스토리 조회
      operationId: getChatHistoryDetail
      tags:
        - 채팅 히스토리
      parameters:
        - name: session_key
          in: path
          description: 세션 키
          required: True
          schema:
            type: string
        - name: user_key
          in: query
          description: 사용자 키
          required: True
          schema:
            type: string
      responses:
        "200":
          description: 성공적으로 특정 채팅 히스토리를 조회했습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryDetailResponse"
        "400":
          description: 요청 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # 공고 목록 조회
  /api/anncs:
    get:
      summary: 공고 목록 조회
      operationId: getAnnouncementList
      tags:
        - 공고
      parameters:
        - name: annc_title
          in: query
          description: 공고 제목 검색 키워드
          required: false
          schema:
            type: string
        - name: annc_status
          in: query
          description: 공고 상태
          required: true
          schema:
            type: string
            enum:
              - 전체
              - 공고중
              - 접수중
              - 접수마감
        - name: annc_type
          in: query
          description: 공고 타입
          required: true
          schema:
            type: string
            enum:
              - 전체
              - 임대
              - 분양
        - name: items_per_page
          in: query
          description: 페이지당 항목 수
          required: true
          schema:
            type: integer
            default: 10
        - name: current_page
          in: query
          description: 현재 페이지 번호
          required: true
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: 성공적으로 공고 목록을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
        "400":
          description: 요청 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # 공고 요약 정보 조회
  /api/annc_summary:
    get:
      summary: 홈 - 공고 요약 데이터 요약
      operationId: getAnnouncementSummary
      tags:
        - 공고 요약
      responses:
        "200":
          description: 성공적으로 공고 요약 정보를 조회했습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnncSummaryResponse"
        "400":
          description: 요청 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      
components:
  schemas:
    # 에러 응답 구조
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: 오류 상세 메시지
          example: 요청 데이터의 유효성 검사에 실패했습니다.
        status:
          type: string
          description: 응답 상태 (error)
          example: error

    # 기본 성공 응답 구조 (Root)
    BaseResponse:
      type: object
      properties:
        message:
          type: string
          description: 요청 처리 결과 메시지
          example: 요청을 성공적으로 처리했습니다.
        status:
          type: string
          description: 응답 상태 (success)
          example: success
        data:
          type: object
          description: 실제 응답 데이터 컨테이너
          nullable: true

    # 공고 목록 응답 스키마
    AnnouncementListResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                page_info:
                  $ref: "#/components/schemas/PageInfo"
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/AnnouncementItem"
                  description: 조회된 공고 항목 리스트

    # 채팅 히스토리 목록 응답 스키마
    ChatHistoriesResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ChatShort"
              description: 조회된 채팅 히스토리 요약 리스트
                  
    # 특정 채팅 히스토리 상세 응답 스키마
    ChatHistoryDetailResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                title:
                  type: string
                  description: 채팅 제목
                  example: 수원 신혼부부 추천 분양
                session_key:
                  type: string
                  description: 세션 ID
                  example: 8a7e0d3c-9b1f-4d2a-8c5e-6f4b3a2d1e0f
                user_key:
                  type: string
                  description: 사용자 키
                  example: 매콤한 숫사슴
                chat_list:
                  type: array
                  description: 채팅 메시지 리스트
                  items:
                    $ref: "#/components/schemas/ChatMessage"

    # 공고 요약 응답 스키마
    AnncSummaryResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                cnt_total:
                  type: integer
                  description: 전체 공고 수
                  example: 1304
                cnt_lease:
                  type: integer
                  description: 임대 공고 수
                  example: 450
                cnt_sale:
                  type: integer
                  description: 분양 공고 수
                  example: 34
                cnt_etc:
                  type: integer
                  description: 기타 공고 수
                  example: 4
                cnt_new_this_week:
                  type: integer
                  description: 이번 주 신규 공고 수
                  example: 12

    # 개별 채팅 메시지 스키마
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
          description: 메시지 ID
          example: 1
        session_id:
          type: string
          description: 세션 ID
          example: 8a7e0d3c-9b1f-4d2a-8c5e-6f4b3a2d1e0f
        sequence:
          type: integer
          description: 메시지 순서
          example: 1
        message_type:
          type: string
          description: 메시지 유형 (system|user|bot)
          example: user
        message:
          type: string
          description: 메시지 내용
          example: 채팅 창에 질문한 내용

    # 페이지네이션 정보 스키마
    PageInfo:
      type: object
      properties:
        total_count:
          type: integer
          description: 전체 항목 수
          example: 35
        current_page:
          type: integer
          description: 현재 페이지 번호
          example: 1
        items_per_page:
          type: integer
          description: 페이지당 항목 수
          example: 10
        total_pages:
          type: integer
          description: 전체 페이지 수
          example: 4
          
    # 공고 개별 항목 스키마
    AnnouncementItem:
      type: object
      properties:
        annc_id:
          type: integer
          description: 공고 ID
          example: 1
        annc_title:
          type: string
          description: 공고 제목
          example: 2025년 상반기 신입사원 공개 채용
        annc_url:
          type: string
          description: 공고 상세 URL
          format: url
          example: https://example.com/announcements/annc_001
        created_at:
          type: string
          description: 생성 일시 (UTC)
          format: date-time
        annc_status:
          type: string
          description: 공고 상태
          example: 공고중
          
    # 채팅 히스토리 요약 스키마
    ChatShort:
      type: object
      properties:
        title: 
          type: string
          description: 채팅 제목
          example: 수원 신혼부부 적합한 공고 알려줘
        session_key:
          type: string
          description: 세션 ID
          example: 8a7e0d3c-9b1f-4d2a-8c5e-6f4b3a2d1e0f

    # 채팅 메시지 요청 스키마
    ChatRequest:
      type: object
      required:
        - user_key
        - session_key
        - user_message
      properties:
        user_key:
          type: string
          description: 사용자 키
          example: 매콤한 숫사슴
        session_key:
          type: string
          description: 세션 키
          example: 8a7e0d3c-9b1f-4d2a-8c5e-6f4b3a2d1e0f
        user_message:
          type: string
          description: 메시지 내용
          example: 수원 신혼부부에게 맞는 공고 찾아줄 수 있어?
        ref_hope_area:
          type: string
          description: 희망지역
          example: 서울, 경기
        ref_age:
          type: integer
          description: 연령, 만 x세
          example: 34세
        ref_marriged:
          type: string
          description: 결혼 여부 (Y/N)
          example: Y
        ref_children:
          type: integer
          description: 자녀 (없음:0, 1:1 ...)
          example: 1
        ref_income:
          type: integer
          description: 연봉(세전) xx 만원
          example: 4500
        ref_annc_id:
          type: integer
          description: 공고 ID
          example: 3

    # 채팅 응답 스키마
    ChatResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                ai_response:
                  type: object
                  description: AI로부터 받은 응답 메시지 정보
                  $ref: "#/components/schemas/ChatMessage"